Clean up signed-unsigned comparison warnings.
git-svn-id: svn://rephial.org/angband/trunk@116 98ddb578-d4ae-11db-a93f-5346e46f259c

+++ src/randname.c
@@ -272,7 +272,7 @@ static void build_prob(name_probs probs, cptr *learn)
  */
 size_t make_word(randname_type name_type, size_t min, size_t max, char *word_buf, size_t buflen)
 {
-	int lnum;
+	size_t lnum;
 	bool found_word = FALSE;
 
 	static name_probs lprobs;
